test() : 기능에 대한 테스트 작성
- 숫자가 아닌 값과 0 이하의 숫자 입력 시 예외 처리
