+++ docs/README.md
@@ -13,15 +13,15 @@
 - [ ] 시도할 횟수를 입력받는 기능
     - [ ] 입력받은 값 만큼 아래 3 가지 프로세스를 반복
 
-##### 1
+##### (1)
 - [ ] 각 자동차 별로 0 ~ 9 의 랜덤한 숫자 뽑기 기능
     - [ ] 독립적인 랜덤 값을 뽑기 위해 자동차 갯수 만큼 진행 
 
-##### 2
+##### (2)
 - [ ] 무작위 값이 4 이상이면 전진, 4 미만이면 제자리인 것을 판단할 기능
     - [ ] 값 결정 후 각 자동차 이름에 해당 값 반영
 
-##### 3
+##### (3)
 - [ ] 각 차수별 결과를 출력하는 기능 
     - [ ] 전진하는 자동차 이름도 같이 출력
     - [ ] 각 자동차가 값 반영 후 얼마나 전진했는지 표시
